New extensions for JTable
Filters extension is far from finished, but its a starting point. Also
included CSS changes for these plugins to work. As Reload and Filters
adds a command button on table title, I needed to float some things to
make it possible to stack buttons to the right of title bar.