Prevent auto reloadTable() on clientOnly AddRecord
When using the clientOnly option for addRecord, the list action would still fire because of a call to _reloadTable() in the _addRow function.  

My change passes the clientOnly option all the way down to _addRecord and performs a check on clientOnly to prevent the _reload from happening.  This allows you to do a manual addRecord using the clientOnly option, which will 1) not create a record on the server and 2) not refresh your table from the server.

The issue was that after adding a manual record using addRecord, your new manually created record disappeared within seconds because the table would automatically reload using your list action (from the server which may not contain your 'clientonly' record).