extend indexing option all the way out to public facing addRecord function
By default addRecord inserts a record row at then end of the list.  internal functions such as _addRow already have the ability to insert the row at any given index (such as index 0 for the top of the list).  This functionality has been extended out to the options list for the public function addRecord.  So given the option 'index: 0', it will place the record at the beginning of the list.  
For example:
     $('#JTableContainer').jtable('addRecord', {
          index: 0,
          record: r,
     });

This will insert a new record at the top of the list, instead of the end.  This is useful when manually adding records to a list that may already be sorted in specific order such as ID DESC.