FEATURE Added Datepicker Year/Month option + getAllRows
Added feature to have default value in options so the user
can select whether the datepicker in jTable forms let's
the user directly change the month and the year.

Added feature to allow user to call public jTable.getAllRows(),
analogous to jTable.getSelectedRows().  Instead of just
returning the selected rows, we now return all the rows
in the jTable (and thus underlying data).

Modified both the jquery.jtable.js source with these changes,
as well as the underlying dev source with identical changes.