Bug fix for checkboxes in form dialogues
This commit fixes a bug in form dialogues where the checkbox property is not actually toggled back a second time when using the label text. The issue is caused by the use of the `checked` attribute (`.attr('checked'`) because changing this attribute is not guarenteed to update the element `checked` property. Attached is an image proving how this can arise in an updated version of Chrome. See how the attribute is present but the property does not match. This was after clicking the `span` element a second time.

Using the `.prop` method of the jQuery we use fixes this issue totally. It is suggested other usage of `attr` on checkboxes is revised.

Proof: http://i.imgur.com/KGM9LUd.png