Add ASP.NET ValidateAntiForgeryToken fix
Jtable ajax calls to controler actions marked with [ValidateAntiForgeryToken]  may not work because Antiforgery checks the request header for an antiforgery token which is then compared with the  cookies token. If they don't match or the token is missing, the server responds with error 400. All this is to prevent Cross-Site Request Forgery (CSRF) Attacks. This fix attempts to manage this by adding a security option where a antiforgerytoken can be passed, this is then added to the ajax call header for server side validation. usage would be like:

 var token = 'gettokenvalue'
        $('#CategoriesContainer').jtable(
            {
                title: 'Categories',
                paging: false,
                sorting: false,
                security:
                {
                    antiforgerytoken: token
                },
                actions: {----------